{boot,reboot}_time: Support the qemu boot loader
When using the qemu boot loader, the kernel command line is set using
the kernel_params configuration setting. This means that a command run
inside the guest can not be used to add the 'S' to the kernel command
line.

Instead, detect this situation and temporarily modify the configuration
parameter.

Signed-off-by: Jonas Eriksson <jonas.eriksson@enea.com>